Refactor GitCaller and GitHelper functions
GitCaller and GitHelper functions are updated to improve readability and functionality. GitCaller's functions are modified to handle argument parameters more efficiently and a new function, FixOctalEncoding, is introduced in GitHelper. This function corrects octal encoding to ensure correct file path is processed. Minor adjustments in function formatting are also made to improve readability.